move failure event emission logic to protocol runner
