+++ modules/cf-core/src/message-handling/handle-protocol-message.ts
@@ -43,14 +43,15 @@ export async function handleReceivedProtocolMessage(
   ]);
   try {
     const { channel } = await protocolRunner.runProtocolWithMessage(
+      router,
       data,
       json && StateChannel.fromJson(json),
     );
     postProtocolStateChannel = channel;
   } catch (error) {
+    // error events are thrown from WITHIN the protocol runner
+    // to ensure they are always emitted symmetrically
     log.error(`Caught error running protocol, aborting. Error: ${error.stack || error.message}`);
-    const outgoingEventData = getOutgoingEventFailureDataFromProtocol(protocol, params!, error);
-    await emitOutgoingMessage(router, outgoingEventData);
     return;
   }
 
@@ -68,62 +69,6 @@ export async function handleReceivedProtocolMessage(
 
 function emitOutgoingMessage(router: RpcRouter, msg: Message) {
   return router.emit(msg["type"], msg, "outgoing");
-}
-
-function getOutgoingEventFailureDataFromProtocol(
-  protocol: ProtocolName,
-  params: ProtocolParam,
-  error: Error,
-): Message {
-  const baseEvent = {
-    from: params.initiatorIdentifier,
-    data: {
-      params,
-      error: error.stack || error.message,
-    },
-  };
-  switch (protocol) {
-    case ProtocolNames.setup: {
-      return {
-        ...baseEvent,
-        type: EventNames.SETUP_FAILED_EVENT,
-      };
-    }
-    case ProtocolNames.sync: {
-      return {
-        ...baseEvent,
-        type: EventNames.SYNC_FAILED_EVENT,
-      };
-    }
-    case ProtocolNames.propose: {
-      return {
-        ...baseEvent,
-        type: EventNames.PROPOSE_INSTALL_FAILED_EVENT,
-      };
-    }
-    case ProtocolNames.install: {
-      return {
-        ...baseEvent,
-        type: EventNames.INSTALL_FAILED_EVENT,
-      };
-    }
-    case ProtocolNames.takeAction: {
-      return {
-        ...baseEvent,
-        type: EventNames.UPDATE_STATE_FAILED_EVENT,
-      };
-    }
-    case ProtocolNames.uninstall: {
-      return {
-        ...baseEvent,
-        type: EventNames.UNINSTALL_FAILED_EVENT,
-      };
-    }
-    default: {
-      const unexpected: never = protocol;
-      throw new Error(`[getOutgoingEventFailureDataFromProtocol] Unexpected case: ${unexpected}`);
-    }
-  }
 }
 
 async function getOutgoingEventDataFromProtocol(
